+++ ports/atmel-samd/common-hal/rgbmatrix/RGBMatrix.c
@@ -26,7 +26,7 @@
 
 #include <stddef.h>
 
-#include "common-hal/protomatter/Protomatter.h"
+#include "common-hal/rgbmatrix/RGBMatrix.h"
 
 #include "samd/timers.h"
 #include "timer_handler.h"
