+++ shared-bindings/rgbmatrix/__init__.c
@@ -29,22 +29,22 @@
 #include "py/obj.h"
 #include "py/runtime.h"
 
-#include "shared-bindings/protomatter/Protomatter.h"
+#include "shared-bindings/rgbmatrix/RGBMatrix.h"
 
-//| :mod:`protomatter` --- Low-level routines for bitbanged LED matrices
+//| :mod:`rgbmatrix` --- Low-level routines for bitbanged LED matrices
 //| =====================================================================
 //|
-//| .. module:: protomatter
+//| .. module:: rgbmatrix
 //|   :synopsis: Low-level routines for bitbanged LED matrices
 //|
 //| .. toctree::
 //|     :maxdepth: 3
 //|
-//|     Protomatter
+//|     RGBMatrix
 
 STATIC const mp_rom_map_elem_t protomatter_module_globals_table[] = {
-    { MP_ROM_QSTR(MP_QSTR___name__), MP_ROM_QSTR(MP_QSTR_protomatter) },
-    { MP_ROM_QSTR(MP_QSTR_Protomatter), MP_ROM_PTR(&protomatter_Protomatter_type) },
+    { MP_ROM_QSTR(MP_QSTR___name__), MP_ROM_QSTR(MP_QSTR_rgbmatrix) },
+    { MP_ROM_QSTR(MP_QSTR_RGBMatrix), MP_ROM_PTR(&protomatter_Protomatter_type) },
 };
 
 STATIC MP_DEFINE_CONST_DICT(protomatter_module_globals, protomatter_module_globals_table);
