+++ shared-module/rgbmatrix/allocator.h
